+++ CHANGELOG.md
@@ -19,6 +19,10 @@ See [Porting Guide](http://docs.ansible.com/ansible/devel/porting_guides/porting
 
 ### New Modules
 
+#### Cloud
+- amazon
+  * aws_caller_facts
+
 
 <a id="2.5"></a>
 
@@ -86,6 +90,7 @@ See [Porting Guide](https://docs.ansible.com/ansible/devel/porting_guides/portin
 * The ANSIBLE_REMOTE_TMP environment variable has been added to supplement (and
   override) ANSIBLE_REMOTE_TEMP.  This matches with the spelling of the config
   value. ANSIBLE_REMOTE_TEMP will be deprecated in the future.
+* A few modules were updated to put temporary files in the existing temp dir already created for the module itself, including get_url, assemble, uri and yum.
 
 #### Removed Modules (previously deprecated):
 * accelerate.